#12 [Feat] 내가 작성한 댓글의 게시물 목록 API 구현

+++ src/main/java/umc/SukBakJi/domain/model/dto/PostListDTO.java
@@ -12,5 +12,5 @@ public class PostListDTO {
     private Long views;
     private String boardName;
     private String menu;
-    private int commentCount;
+    private long commentCount;
 }
